+++ renderer/src/data/changelog.js
@@ -1,22 +1,11 @@
 export default {
-    description: "Expect more soon!",
+    description: "Improvements coming soon!",
     changes: [
-        {
-            title: "What's new?",
-            type: "added",
-            items: [
-                "New notices API for developers available via `BdApi.showNotice`! Docs coming soon."
-            ]
-        },
         {
             title: "Fixes",
             type: "fixed",
             items: [
-                "Outdated styles and classnames have been fixed.",
-                "Public servers button shows again.",
-                "Disable Media Keys option actually prevents Discord from hijacking media keys again.",
-                "CustomCSS is removed if you disable the system.",
-                "CSS editor popout no longer turns the screen gray."
+                "BD loads again, thanks to Strencher for this late night fix while I was away"
             ]
         }
     ]