+++ renderer/src/data/settings.js
@@ -35,16 +35,16 @@ export default [
     },
     {
         type: "category",
-        id: "developer",
+        id: "editor",
         collapsible: true,
         shown: false,
         settings: [
-            {type: "switch", id: "debugLogs", value: false},
-            {type: "switch", id: "devTools", value: false},
-            {type: "switch", id: "debuggerHotkey", value: false, enableWith: "devTools"},
-            {type: "switch", id: "reactDevTools", value: false, enableWith: "devTools"},
-            {type: "switch", id: "inspectElement", value: false, enableWith: "devTools"},
-            {type: "switch", id: "devToolsWarning", value: false, enableWith: "devTools"},
+            {type: "switch", id: "lineNumbers", value: true},
+            {type: "switch", id: "minimap", value: true},
+            {type: "switch", id: "hover", value: true},
+            {type: "switch", id: "quickSuggestions", value: true},
+            {type: "number", id: "fontSize", min: 2, value: 14},
+            {type: "dropdown", id: "renderWhitespace", value: "selection", options: [{value: "none"}, {value: "all"}, {value: "selection"}]}
         ]
     },
     {
@@ -56,6 +56,20 @@ export default [
             {type: "switch", id: "transparency", value: false},
             {type: "switch", id: "removeMinimumSize", value: false},
             {type: "switch", id: "frame", value: false, hidden: true}
+        ]
+    },
+    {
+        type: "category",
+        id: "developer",
+        collapsible: true,
+        shown: false,
+        settings: [
+            {type: "switch", id: "debugLogs", value: false},
+            {type: "switch", id: "devTools", value: false},
+            {type: "switch", id: "debuggerHotkey", value: false, enableWith: "devTools"},
+            {type: "switch", id: "reactDevTools", value: false, enableWith: "devTools"},
+            {type: "switch", id: "inspectElement", value: false, enableWith: "devTools"},
+            {type: "switch", id: "devToolsWarning", value: false, enableWith: "devTools"},
         ]
     }
 ];